[attestation][amd] Support for extracting PSP header from key database
After parsing AMD public key, the next item to parse and validate
is the key database. That contains additional keys that might be used
for other blobs in the firmware. The key database, as other blobs on the
platform, is pre-pended with a header and appended with a signature. This
parsing logic should be the same for PSP off chip bootloader, SMU FW, etc.

This diff specifically adds support for parsing the header structure for
Key Database and validating its signature against AMD Public key.