[attestation][amd] Decide signature size based on key parameters
In the extraction of the key db signature, we were making the
assumption the signature came from RSA 4096b key. This is probably not
going to change, as the keydb is signed by AMD root key and that should
remain a RSA 4096b key. However, this check should not be hardcoded,
and we should instead match the key id in the PSP blob to the content
of the key database, which should be pre-populated with AMD public key.