Change MeasurementIDDXE to EV_EFI_PLATFORM_FIRMWARE_BLOB2
This supports having multiple types for measurements. Right now,
this will only be used for EV_EFI_PLATFORM_FIRMWARE_BLOB2 which is
defined in the newer TCG spec.

It's possible that we can make this more intelligent, by examining
a certain section of the firmware to determine the TCG spec used.

However, I do not know of such a section.

The caveat to this approach is that it will choose whatever matches
first in the slice. Right now with EV_EFI_PLATFORM_FIRMWARE_BLOB2 this
will not be an issue, since EV_POST_CODE is not used when
EV_EFI_PLATFORM_FIRMWARE_BLOB2 is used.

In the future, this approach may need to be refined

Signed-off-by: Trynity Mirell <trynity@fb.com>