Add a skeleton of redesigned "pcr" package
Adding the main entities of a new design of PCR values calculations
process. New package is called "bootflow" (instead of "pcr"), since
it is not longer focused on PCR values only. It also supports other
roots of trust for storage and measurements.

New design (in contrast to the old one) is:
* Extensible from external packages.
* TPM terms "Event", "Extend" and "Measurement" are clarified.
* Flow now consists of steps. Each step may contain multiple
  measurements. It solves the problem with dynamic amount of
  measurements on some platforms.
* New design also aware of DICE and other chains of trust.